# ch/inss/jwserver/openapi/openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: john.doe@example.com
    name: John Doe
    url: http://example.com
  description: OpenAPI specification generated by Joaswizard (https://github.com/networkinss/joaswizard).
  license:
    name: Apache 2.0 license
    url: http://opensource.org/licenses/Apache-2.0
  title: "Task, metadata, shell, resources API"
  version: 1.0.0
servers:
- description: Local environment.
  url: http://localhost:8080/api/v1
paths:
  /metadata:
    post:
      description: Creates new metadata.
      operationId: add_metadata
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
        description: Metadata to be created
        required: true
        x-name: metadata
      responses:
        "201":
          description: metadata created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Creates new metadata
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /metadata/{name}:
    delete:
      description: Deletes metadata by name.
      operationId: delete_metadata
      parameters:
      - description: ID of metadata to delete
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: metadata deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Deletes metadata by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    get:
      description: Returns metadata by name.
      operationId: find_metadata_byname
      parameters:
      - description: name of metadata to fetch
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Metadata'
          description: Metadata returned
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Returns metadata by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    put:
      description: Updates metadata by name.
      operationId: update_metadata
      parameters:
      - description: ID of metadata to put
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Metadata'
        description: Metadata to be updated
        required: true
        x-name: metadata
      responses:
        "204":
          description: metadata updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Updates metadata by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /metadatas:
    get:
      description: List of metadatas.
      operationId: get_metadata_list
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Metadata'
                type: array
          description: Metadata List
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: List of metadatas
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /resources:
    post:
      description: Creates new resources.
      operationId: add_resources
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/addResources_request'
      responses:
        "201":
          description: resource file created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Creates new resources
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /resources/{name}:
    delete:
      description: Deletes resources by name.
      operationId: delete_resources
      parameters:
      - description: ID of resources to delete
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: resources deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Deletes resources by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    get:
      description: Returns resources by name.
      operationId: find_resources_byname
      parameters:
      - description: name of resources to fetch
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resources'
          description: Resources returned
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Returns resources by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    put:
      description: Updates resources by name.
      operationId: update_resources
      parameters:
      - description: ID of resources to put
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Resources'
        description: Resources to be updated
        required: true
        x-name: resources
      responses:
        "204":
          description: resources updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Updates resources by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /resourcess:
    get:
      description: List of resourcess.
      operationId: get_resources_list
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Resources'
                type: array
          description: Resources List
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: List of resourcess
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /shell:
    post:
      description: Creates new shell.
      operationId: add_shell
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/addShell_request'
      responses:
        "201":
          description: shellscript created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Creates new shell
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /shell/{name}:
    delete:
      description: Deletes shell by name.
      operationId: delete_shell
      parameters:
      - description: ID of shell to delete
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: shell deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Deletes shell by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    get:
      description: Returns shell by name.
      operationId: find_shell_byname
      parameters:
      - description: name of shell to fetch
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shell'
          description: Shell returned
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Returns shell by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    put:
      description: Updates shell by name.
      operationId: update_shell
      parameters:
      - description: ID of shell to put
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Shell'
        description: Shell to be updated
        required: true
        x-name: shell
      responses:
        "204":
          description: shell updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Updates shell by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /shells:
    get:
      description: List of shells.
      operationId: get_shell_list
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Shell'
                type: array
          description: Shell List
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: List of shells
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /task:
    post:
      description: Creates new task.
      operationId: add_task
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        description: Task to be created
        required: true
        x-name: task
      responses:
        "201":
          description: task created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Creates new task
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /task/{name}:
    delete:
      description: Deletes task by name.
      operationId: delete_task
      parameters:
      - description: ID of task to delete
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: task deleted
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Deletes task by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    get:
      description: Returns task by name.
      operationId: find_task_byname
      parameters:
      - description: name of task to fetch
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
          description: Task returned
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Returns task by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
    put:
      description: Updates task by name.
      operationId: update_task
      parameters:
      - description: ID of task to put
        explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
        description: Task to be updated
        required: true
        x-name: task
      responses:
        "204":
          description: task updated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Data with the specified ID was not found.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: Updates task by name
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
  /tasks:
    get:
      description: List of tasks.
      operationId: get_task_list
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Task'
                type: array
          description: Task List
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Bad request.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Authorization information is missing or invalid.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
          description: Unexpected internal server error.
      summary: List of tasks
      x-openapi-router-controller: ch.inss.jwserver.api.default_controller
components:
  schemas:
    ErrorModel:
      properties:
        code:
          format: int32
          title: code
          type: integer
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: ErrorModel
      type: object
...
    Metadata:
      example:
        binaryfiles:
        - file1
        - file2
        - file3
        appname: demoapp
        author: John
        domain: example.com
        dbtype: postgres
        username: johndoe
      properties:
        author:
          example: John
          minLength: 4
          title: author
          type: string
        username:
          example: johndoe
          minLength: 7
          title: username
          type: string
        appname:
          example: demoapp
          minLength: 7
          title: appname
          type: string
        domain:
          example: example.com
          minLength: 11
          title: domain
          type: string
        dbtype:
          example: postgres
          minLength: 8
          title: dbtype
          type: string
        binaryfiles:
          example:
          - file1
          - file2
          - file3
          items:
            type: string
          minLength: 5
          title: binaryfiles
          type: array
      required:
      - appname
      - author
      - binaryfiles
      - dbtype
      - domain
      - username
      title: Metadata
      type: object
    Shell:
      example:
        filename: shellscript.sh
      properties:
        filename:
          example: shellscript.sh
          minLength: 14
          title: filename
          type: string
      required:
      - filename
      title: Shell
      type: object
    Resources:
      example:
        filename: example.mustache
      properties:
        filename:
          example: example.mustache
          minLength: 16
          title: filename
          type: string
      required:
      - filename
      title: Resources
      type: object
    addShell_request:
      properties:
        filestream:
          format: binary
          type: string
        filename:
          description: Any shell script to execute a task.
          example: starttaskgenerate.sh
          format: string
          type: string
      required:
      - filestream
      type: object
    addResources_request:
      properties:
        filestream:
          format: binary
          type: string
        filename:
          description: Any file that is needed by a task. Can be a mustache file.
          example: sqlc.mustache
          format: string
          type: string
      required:
      - filestream
      type: object